feat: 스타일 카테고리 리스트 fetch
