+++ src/app/api/auth/[...nextauth]/options.ts
@@ -42,9 +42,23 @@ export const options: NextAuthOptions = {
       clientSecret: process.env.KAKAO_CLIENT_SECERET || '',
     }),
   ],
+
   callbacks: {
     async jwt({ token, user }) {
-      return { ...token, ...user }
+      // 유효시간 6시간으로 설정
+      if (user) {
+        // 첫 로그인시에만 user가 존재
+        return { ...token, ...user, expiresAt: Date.now() + 60 * 60 * 6 * 1000 }
+      }
+      if (Date.now() < (token.expiresAt as number)) {
+        // 첫 로그인 이후 token access
+        // console.log(Date.now(), token.expiresAt as number)
+        return token
+      }
+      // 유효시간 만료시 refresh token으로 재발급하는 로직 추가
+      console.log('error: 유효시간 만료')
+
+      return token
     },
     async session({ session, token }) {
       const updatedSession = {
@@ -55,6 +69,7 @@ export const options: NextAuthOptions = {
           refreshToken: `${token.refreshToken}`,
         },
       }
+
       return updatedSession
     },
   },
