+++ src/components/pages/partner/mypage/profile/PartnerChatTime.tsx
@@ -7,7 +7,7 @@ import StretchedRoundedButton from '@/components/ui/button/StretchedRoundedButto
 import TimepickerSwiper from '@/components/ui/picker/TimepickerSwiper'
 import useToast from '@/stores/toast'
 import { formatTime } from '@/utils/formatTime'
-import { updateChatTime } from '@/app/api/partner/PartnerProfile'
+import { updateChatTime } from '@/app/api/partner/PartnerProfileUpdate'
 
 export default function PartnerChatTime({
   startChatTime,
