+++ src/components/pages/partner/mypage/profile/ProfileImage.tsx
@@ -8,7 +8,7 @@ import {
   deleteAndUpdateImage,
 } from '@/utils/uploadImage'
 import SliderModal from '@/components/common/SliderModal'
-import { updatePartnerProfileImage } from '@/app/api/partner/PartnerProfile'
+import { updatePartnerProfileImage } from '@/app/api/partner/PartnerProfileUpdate'
 
 export default function ProfileImage({
   profileImage,
