+++ src/components/pages/partner/mypage/profile/career/PartnerCareerDashboard.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import Image from 'next/image'
 import { PartnerCareerFetchType } from '@/types/partnerProfileTypes'
 import { formatCareerPeriod } from '@/utils/formatTime'
-import { deletePartnerCareer } from '@/app/api/partner/PartnerProfile'
+import { deletePartnerCareer } from '@/app/api/partner/PartnerProfileUpdate'
 import PartnerCareerEditor from '@/components/pages/partner/mypage/profile/career/PartnerCareerEditor'
 import useConfirmStore from '@/stores/confirm'
 import useModal from '@/stores/modal'
