--- src/libs/memberFavoritStyleData.ts
@@ -1,78 +1,0 @@
-import { MemberStyleType } from '@/types/commonTypes'
-
-const styleInfo: MemberStyleType[] = [
-  {
-    id: 1,
-    value: 'casual',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 2,
-    value: 'street',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 3,
-    value: 'vintage',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 4,
-    value: 'normcore',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 5,
-    value: 'modern',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 6,
-    value: 'feminine',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 7,
-    value: 'dandy',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 8,
-    value: 'sporty',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 9,
-    value: 'classic',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 10,
-    value: 'amekaji',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 11,
-    value: 'highteen',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-  {
-    id: 12,
-    value: 'minimal',
-    image:
-      'https://personull-bucket.s3.ap-northeast-2.amazonaws.com/favorite-style/amekaji.png',
-  },
-]
-
-export default styleInfo
