+++ src/types/baseResponseType.ts
@@ -1,0 +1,7 @@
+export interface BaseResponseType {
+  httpStatus: string
+  isSuccess: boolean
+  message: string
+  code: number
+  result: string[] | object[] | number[]
+}
